renamed and small fixes
